+++ CHANGELOG.md
@@ -11,6 +11,26 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 \-
 
+## [1.0.0](https://github.com/xxlabaza/median-meter/releases/tag/1.0.0) - 2019-01-15
+
+Add Kubernetes support.
+
+### Added
+
+- Write Kubernetes deploy file;
+- Add `LeaderService` to Hazelcast cluster, which is responsible for calling different handlers on becoming leader/regular member;
+- Complex (discovery, hazelcast, median logic) test;
+- `pom.xml` parameter `skipAllTests` - for skipping all tests and checking, for example, `PMD`, `spotbugs` or `checkstyle`.
+
+### Changed
+
+- Separate CLI-arguments parsing and app launching;
+- `MqttClientWrapper` has a caching static method for building a client;
+- `MqttClientWrapper` has a registry of all subscriptions;
+- Don't disconnect from `MQTT`, when `MqttDiscoveryServiceClient` is stopped;
+- Add more debug logs;
+- Properly stop/close services (add null-checks);
+
 ## [0.3.0](https://github.com/xxlabaza/median-meter/releases/tag/0.3.0) - 2019-01-15
 
 Create median meter function.